Removed `Carry` state. Merged it with Carry State.
