--- Assets/Scripts/NPCs/States/CarryState.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cc8515c44c3500849a0673690c393e0e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
